test: consolidate fragmented checkpoint integration tests into a single file
